refacto mnist and unet
